Introducing Saving & Just Print!
Model is now saved to a .tflearn file after training.
Also added stripped down main called, Just_Generate.py, that skips all the training formalities, grabs the latest checkpoint and prints output based on it. It does not use the .tflearn file, only because checkpoints tend to be more up to date. (that might change later)